+++ src/Matrix.h
@@ -10,21 +10,21 @@ private:
     // On considère une matrice composée de vecteurs colonnes
     std::vector<Vector> m_matrix;
 
-    size_t m_rows; // = m_dimension sur tous les Vector de la Matrix
-    size_t m_cols;
+    int m_rows; // = m_dimension sur tous les Vector de la Matrix
+    int m_cols;
 public:
-    Matrix( const size_t& nbRows, const size_t& nbCols );
+    Matrix( const int& nbRows, const int& nbCols );
     Matrix( const Matrix& matrix );
 
     // Getters
-    const size_t& getNbRows() const { return m_rows; }
-    const size_t& getNbColumns() const { return m_cols; }
+    const int& getNbRows() const { return m_rows; }
+    const int& getNbColumns() const { return m_cols; }
 
     // Setters
     void setMatrix( float values[] );
 
-    const Vector& operator[]( const size_t& vectorIndex ) const { return m_matrix[ vectorIndex ]; }
-    Vector& operator[]( const size_t& vectorIndex ) { return m_matrix[ vectorIndex ]; }
+    const Vector& operator[]( const int& vectorIndex ) const { return m_matrix[ vectorIndex ]; }
+    Vector& operator[]( const int& vectorIndex ) { return m_matrix[ vectorIndex ]; }
     Matrix operator*( const float& value ) const;
     Matrix operator*( const Matrix& matrix ) const;
